feat: Implement authentication checks for various API routes to ensure user authorization

+++ app/api/delete/delete-post/route.ts
@@ -1,8 +1,16 @@
 import prisma from '@/lib/db';
+import { auth } from '@/auth';
 import { NextResponse } from 'next/server';
 
 export async function DELETE(request: Request) {
   try {
+    const session = await auth();
+    if (!session || !session.user) {
+      return NextResponse.json('Unauthorized', {
+        status: 401,
+      });
+    }
+
     const { searchParams } = new URL(request.url);
     const id = searchParams.get('id');
 
@@ -10,6 +18,19 @@ export async function DELETE(request: Request) {
       return NextResponse.json('ID of the post is required', {
         status: 400,
       });
+    }
+
+    const post = await prisma.posts.findUnique({
+      where: { id },
+      select: { authorId: true },
+    });
+
+    if (!post) {
+      return NextResponse.json('Post not found', { status: 404 });
+    }
+
+    if (post.authorId !== session.user.id && session.user.role !== 'ADMIN') {
+      return NextResponse.json('Forbidden', { status: 403 });
     }
 
     const deletedPost = await prisma.posts.delete({
